Move HOM strategies to another file
