+++ src/pynguin/assertion/mutation_analysis/controller.py
@@ -12,14 +12,13 @@ from __future__ import annotations
 
 import ast
 import inspect
-import random
 import types
 
-from typing import Generator, Callable
+from typing import Generator
 
 from pynguin.assertion.mutation_analysis import utils
 from pynguin.assertion.mutation_analysis.operators.base import Mutation, MutationOperator
-
+from pynguin.assertion.mutation_analysis.stategies import HOMStrategy, FirstToLastHOMStrategy
 
 class MutationController:
 
@@ -49,102 +48,6 @@ class MutationController:
             )
         except BaseException:
             return None
-
-
-class HOMStrategy:
-
-    def __init__(self, order: int = 2) -> None:
-        self.order = order
-
-    def remove_bad_mutations(
-        self,
-        mutations_to_apply: list[Mutation],
-        available_mutations: list[Mutation],
-        allow_same_operators: bool = True
-    ) -> None:
-        for mutation_to_apply in mutations_to_apply:
-            for available_mutation in available_mutations.copy():
-                if (
-                    mutation_to_apply.node == available_mutation.node
-                    or mutation_to_apply.node in getattr(available_mutation.node, "children")
-                    or available_mutation.node in getattr(mutation_to_apply.node, "children")
-                    or (
-                        not allow_same_operators
-                        and mutation_to_apply.operator == available_mutation.operator
-                    )
-                ):
-                    available_mutations.remove(available_mutation)
-
-
-class FirstToLastHOMStrategy(HOMStrategy):
-
-    def generate(self, mutations: list[Mutation]) -> Generator[list[Mutation], None, None]:
-        mutations = mutations.copy()
-        while mutations:
-            mutations_to_apply: list[Mutation] = []
-            index = 0
-            available_mutations = mutations.copy()
-            while len(mutations_to_apply) < self.order and available_mutations:
-                mutation = available_mutations.pop(index)
-                mutations_to_apply.append(mutation)
-                mutations.remove(mutation)
-                index = 0 if index == -1 else -1
-                self.remove_bad_mutations(mutations_to_apply, available_mutations)
-            yield mutations_to_apply
-
-
-class EachChoiceHOMStrategy(HOMStrategy):
-
-    def generate(self, mutations: list[Mutation]) -> Generator[list[Mutation], None, None]:
-        mutations = mutations.copy()
-        while mutations:
-            mutations_to_apply: list[Mutation] = []
-            available_mutations = mutations.copy()
-            while len(mutations_to_apply) < self.order and available_mutations:
-                mutation = available_mutations.pop(0)
-                mutations_to_apply.append(mutation)
-                mutations.remove(mutation)
-                self.remove_bad_mutations(mutations_to_apply, available_mutations)
-            yield mutations_to_apply
-
-
-class BetweenOperatorsHOMStrategy(HOMStrategy):
-
-    def generate(self, mutations: list[Mutation]) -> Generator[list[Mutation], None, None]:
-        usage = {mutation: 0 for mutation in mutations}
-        not_used = mutations.copy()
-        while not_used:
-            mutations_to_apply: list[Mutation] = []
-            available_mutations = mutations.copy()
-            available_mutations.sort(key=lambda x: usage[x])
-            while len(mutations_to_apply) < self.order and available_mutations:
-                mutation = available_mutations.pop(0)
-                mutations_to_apply.append(mutation)
-                if usage[mutation] == 0:
-                    not_used.remove(mutation)
-                usage[mutation] += 1
-                self.remove_bad_mutations(mutations_to_apply, available_mutations, allow_same_operators=False)
-            yield mutations_to_apply
-
-
-class RandomHOMStrategy(HOMStrategy):
-
-    def __init__(self, order: int = 2, shuffler: Callable = random.shuffle) -> None:
-        super().__init__(order)
-        self.shuffler = shuffler
-
-    def generate(self, mutations: list[Mutation]) -> Generator[list[Mutation], None, None]:
-        mutations = mutations.copy()
-        self.shuffler(mutations)
-        while mutations:
-            mutations_to_apply: list[Mutation] = []
-            available_mutations = mutations.copy()
-            while len(mutations_to_apply) < self.order and available_mutations:
-                mutation = available_mutations.pop(0)
-                mutations_to_apply.append(mutation)
-                mutations.remove(mutation)
-                self.remove_bad_mutations(mutations_to_apply, available_mutations)
-            yield mutations_to_apply
 
 
 class FirstOrderMutator:
